refactor(Layout): port typescript + convert to functional component
